# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ingredient {
  name: string;
  quantity: string;
  unit: string;
  unitCost: number;
}

export interface Supply {
  name: string;
  quantity: string;
  unitCost: number;
}

export interface Dish {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  recipe?: string | null;
  /** @nullable */
  prep?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  flatware?: string | null;
  ingredients: Ingredient[];
  supplies: Supply[];
  /**
   * e.g. appetizer, main, dessert, side
   * @nullable
   */
  category?: string | null;
  createdAt: string;
}

export interface CreateDishBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  recipe?: string | null;
  /** @nullable */
  prep?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  flatware?: string | null;
  ingredients: Ingredient[];
  supplies: Supply[];
  /** @nullable */
  category?: string | null;
}

export interface UpdateDishBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  recipe?: string | null;
  /** @nullable */
  prep?: string | null;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  flatware?: string | null;
  ingredients?: Ingredient[];
  supplies?: Supply[];
  /** @nullable */
  category?: string | null;
}

export type EventRequestStatus =
  (typeof EventRequestStatus)[keyof typeof EventRequestStatus];

export const EventRequestStatus = {
  new: "new",
  in_progress: "in_progress",
  confirmed: "confirmed",
} as const;

export interface EventRequest {
  id: number;
  clientName: string;
  clientEmail: string;
  eventDate: string;
  guestCount: number;
  eventType: string;
  /** @nullable */
  restrictions?: string | null;
  /** @nullable */
  notes?: string | null;
  status: EventRequestStatus;
  createdAt: string;
}

export interface CreateEventBody {
  clientName: string;
  clientEmail: string;
  eventDate: string;
  guestCount: number;
  eventType: string;
  /** @nullable */
  restrictions?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type UpdateEventBodyStatus =
  (typeof UpdateEventBodyStatus)[keyof typeof UpdateEventBodyStatus];

export const UpdateEventBodyStatus = {
  new: "new",
  in_progress: "in_progress",
  confirmed: "confirmed",
} as const;

export interface UpdateEventBody {
  status?: UpdateEventBodyStatus;
  clientName?: string;
  clientEmail?: string;
  eventDate?: string;
  guestCount?: number;
  eventType?: string;
  /** @nullable */
  restrictions?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface MenuEntry {
  id: number;
  eventId: number;
  dishId: number;
  /** e.g. appetizer, main, dessert, side, beverage */
  course: string;
  sortOrder: number;
  dish: Dish;
}

export interface MenuEntryInput {
  dishId: number;
  course: string;
  sortOrder: number;
}

export interface SetEventMenuBody {
  entries: MenuEntryInput[];
}

export type CostLineType = (typeof CostLineType)[keyof typeof CostLineType];

export const CostLineType = {
  ingredient: "ingredient",
  supply: "supply",
} as const;

export interface CostLine {
  name: string;
  quantity: string;
  unitCost: number;
  totalCost: number;
  type: CostLineType;
}

export interface EventCost {
  eventId: number;
  guestCount: number;
  atCost: number;
  margin: number;
  totalPrice: number;
  lines: CostLine[];
}

export interface IngredientAggregate {
  name: string;
  unit: string;
  totalQuantity: number;
  totalCost: number;
}

export interface SupplyAggregate {
  name: string;
  totalQuantity: number;
  totalCost: number;
}

export interface DashboardSummary {
  totalEvents: number;
  newEvents: number;
  inProgressEvents: number;
  confirmedEvents: number;
  totalDishes: number;
  totalEstimatedRevenue: number;
  totalAtCost: number;
  upcomingEvents: EventRequest[];
  ingredientTotals: IngredientAggregate[];
  supplyTotals: SupplyAggregate[];
  flatwareRequired: string[];
}

export type ProcurementItemType =
  (typeof ProcurementItemType)[keyof typeof ProcurementItemType];

export const ProcurementItemType = {
  ingredient: "ingredient",
  supply: "supply",
} as const;

export interface ProcurementItem {
  name: string;
  totalQuantity: string;
  unit: string;
  estimatedCost: number;
  type: ProcurementItemType;
  events: string[];
}

export type GetEventsParams = {
  status?: GetEventsStatus;
};

export type GetEventsStatus =
  (typeof GetEventsStatus)[keyof typeof GetEventsStatus];

export const GetEventsStatus = {
  new: "new",
  in_progress: "in_progress",
  confirmed: "confirmed",
} as const;

export type GetEventCostParams = {
  /**
   * Margin multiplier (e.g. 1.3 = 30% margin). Defaults to 1.
   */
  margin?: number;
};

export type GetProcurementListParams = {
  eventId?: number;
};